Explore data (#2)
* Divided scripts into sub-folders

Explore data contains scripts used to analyze data; query scripts contains those designed to query data using R and SQL; shiny applications contains plots and tables written in shiny and wrapped in plotly; Validation scripts are those which are used to confirm data and/or identify issues.

* Update data_summary.R

Update the formula for sample standard deviation, added instruction on how to use the function on a list of data frames, and fixed the date time variables so that the min and max fields do not convert the values and treat them like dates (grabbing the earliest and latest).

* data summary example

I adjusted the fake data used in the sample and added some instructions to make the script clearer. Also added a reminder not to install the moments library if not already installed.

* SQL

Added attempts at compiling SQL

* Create eda_correlations.R

Script that uses the easy stats package to quickly test and plot different types of correlations (includes multi-level and grouped)

* Create one_way_ANOVA.R

Created script for a one-way ANOVA loop. For the future versions I need to account for family-wise error, include effect size, and drop unneeded columns.