Merged to main (#10)
* Create SECURITY.md

* Explore data (#8)

* Divided scripts into sub-folders

Explore data contains scripts used to analyze data; query scripts contains those designed to query data using R and SQL; shiny applications contains plots and tables written in shiny and wrapped in plotly; Validation scripts are those which are used to confirm data and/or identify issues.

* Update data_summary.R

Update the formula for sample standard deviation, added instruction on how to use the function on a list of data frames, and fixed the date time variables so that the min and max fields do not convert the values and treat them like dates (grabbing the earliest and latest).

* data summary example

I adjusted the fake data used in the sample and added some instructions to make the script clearer. Also added a reminder not to install the moments library if not already installed.

* SQL

Added attempts at compiling SQL

* Create eda_correlations.R

Script that uses the easy stats package to quickly test and plot different types of correlations (includes multi-level and grouped)

* Create one_way_ANOVA.R

Created script for a one-way ANOVA loop. For the future versions I need to account for family-wise error, include effect size, and drop unneeded columns.

* Update one_way_ANOVA.R

Removed unnecessary variables

* Create excel_splitter.R

Function designed to split a large excel file into multiple sheets based on the number of rows per sheet desired.

* Added the shiny correlation app

Update the eda application script with a source to direct back to for the shiny app. Also added the script where the walkthrough was made into a shiny applications which can be used for analysis.

* Application push

Additional applications for analysis

* Update eda_correlations.R

* Create 08_shiny_scatter_parameters_flex.R

Update data explorer scatter plot application

* Add app for missing data

Added app for detecting missing data

* Likert explorer

Add application for exploring likert scales

* Adjust names

For easier sorting and pulling

* Added t-sql descriptives

Scripts which calculate descriptives across a table using T-SQL

* Update applications and remove history files

Removed R history and added function for rescaled scatter explorer and application in shiny/plotly

* Added versions without plotly

And moved the plotly versions into a sub folder

* Update

* Explore data (#9)

* Divided scripts into sub-folders

Explore data contains scripts used to analyze data; query scripts contains those designed to query data using R and SQL; shiny applications contains plots and tables written in shiny and wrapped in plotly; Validation scripts are those which are used to confirm data and/or identify issues.

* Update data_summary.R

Update the formula for sample standard deviation, added instruction on how to use the function on a list of data frames, and fixed the date time variables so that the min and max fields do not convert the values and treat them like dates (grabbing the earliest and latest).

* data summary example

I adjusted the fake data used in the sample and added some instructions to make the script clearer. Also added a reminder not to install the moments library if not already installed.

* SQL

Added attempts at compiling SQL

* Create eda_correlations.R

Script that uses the easy stats package to quickly test and plot different types of correlations (includes multi-level and grouped)

* Create one_way_ANOVA.R

Created script for a one-way ANOVA loop. For the future versions I need to account for family-wise error, include effect size, and drop unneeded columns.

* Update one_way_ANOVA.R

Removed unnecessary variables

* Create excel_splitter.R

Function designed to split a large excel file into multiple sheets based on the number of rows per sheet desired.

* Added the shiny correlation app

Update the eda application script with a source to direct back to for the shiny app. Also added the script where the walkthrough was made into a shiny applications which can be used for analysis.

* Application push

Additional applications for analysis

* Update eda_correlations.R

* Create 08_shiny_scatter_parameters_flex.R

Update data explorer scatter plot application

* Add app for missing data

Added app for detecting missing data

* Likert explorer

Add application for exploring likert scales

* Adjust names

For easier sorting and pulling

* Added t-sql descriptives

Scripts which calculate descriptives across a table using T-SQL

* Update applications and remove history files

Removed R history and added function for rescaled scatter explorer and application in shiny/plotly

* Added versions without plotly

And moved the plotly versions into a sub folder

* Update

* Regression applications

Added a few of the regression applications. The base has a scatter plot and model results, while the second one includes a loess selection for viewing a potential fit. The third includes a line of best fit for a linear model to compare.

* Regression

Added a script with plotly on the scatter plot and added packages to script 3.