Formatting phone numbers
Have an R function and SQL script to validate and fix phone numbers. Both put the numbers in (555)-555-5555 if 10 numbers and 555-5555 if 7 numbers, and add a column with a '1' if there are fewer than 7 or 10 numbers. The SQL script first cleans the phone column (in the original table) and then tries to format. The R function also removes all non-numbers from the column and uses and ifelse to determine formatting and rules for the number flag column.